+++ common/src/main/scala/org/dka/rdbms/common/dao/PublisherDao.scala
@@ -1,0 +1,10 @@
+package org.dka.rdbms.common.dao
+
+import org.dka.rdbms.common.model.{ID, Publisher}
+
+/**
+ * adds methods beyond simple crud stuff anticipated to be mostly specific queries
+ *
+ * this interface is db agnostic and allows for easy unit testing since an database is not required
+ */
+trait PublisherDao extends CrudDao[Publisher, ID] {}
