+++ common/src/main/scala/org/dka/rdbms/common/model/City.scala
@@ -1,0 +1,18 @@
+package org.dka.rdbms.common.model
+
+import io.circe._
+
+final case class City(value: String) extends StringItem {
+  override val fieldName: String = City.fieldName
+}
+
+object City {
+  val fieldName: String = "city"
+  def apply(o: Option[String]): Option[City] = o.map(City(_))
+  def toJsonLine(item: City): (String, Json) = (fieldName, Json.fromString(item.value))
+  def toJsonLine(item: Option[City]): Option[(String, Json)] = item.map(toJsonLine)
+
+  def fromJsonLine(c: HCursor): Either[DecodingFailure, City] = StringItem.fromJsonLine(c, fieldName)(City.apply)
+  def fromOptionalJsonLine(c: HCursor): Either[DecodingFailure, Option[City]] =
+    StringItem.fromOptionalJsonLine(c, fieldName)(apply)
+}
