+++ common/src/main/scala/org/dka/rdbms/common/model/RoyaltySchedule.scala
@@ -1,0 +1,9 @@
+package org.dka.rdbms.common.model
+
+import io.circe._
+
+final case class RoyaltySchedule(
+  titleId: String,
+  lowRange: Int,
+  highRange: Int,
+  royalty: BigDecimal)
