--- common/src/main/scala/org/dka/rdbms/model/RoyaltySchedule.scala
@@ -1,9 +1,0 @@
-package org.dka.rdbms.model
-
-import io.circe._
-
-final case class RoyaltySchedule(
-  titleId: String,
-  lowRange: Int,
-  highRange: Int,
-  royalty: BigDecimal)
