--- common/src/main/scala/org/dka/rdbms/model/dao/PublisherDao.scala
@@ -1,10 +1,0 @@
-package org.dka.rdbms.model.dao
-
-import org.dka.rdbms.model.{ID, Publisher}
-
-/**
- * adds methods beyond simple crud stuff anticipated to be mostly specific queries
- *
- * this interface is db agnostic and allows for easy unit testing since an database is not required
- */
-trait PublisherDao extends CrudDao[Publisher, ID] {}
