+++ slick/src/main/scala/org/dka/rdbms/slick/dao/DaoFactory.scala
@@ -1,0 +1,50 @@
+package org.dka.rdbms.slick.dao
+
+import com.typesafe.config.ConfigFactory
+import pureconfig.ConfigSource
+import pureconfig.generic.auto._
+import slick.jdbc.JdbcBackend.Database
+import com.typesafe.scalalogging.Logger
+import org.dka.rdbms.common.dao.{AuthorDao, ConfigurationException, PublisherDao}
+import org.dka.rdbms.slick.config.DBConfig
+import org.dka.rdbms.slick.config.DBConfig._  // must be kept even though intellij complains
+
+import scala.util.{Failure, Try} // must be kept even though intellij thinks it is unused
+
+class DaoFactory(val database: Database) {
+  val authorsDao: AuthorDao = new AuthorDaoImpl(database)
+  val publisherDao: PublisherDao = new PublisherDaoImpl(database)
+}
+
+object DaoFactoryBuilder {
+  private val logger = Logger(getClass.getName)
+  lazy val configure: Either[ConfigurationException, DaoFactory] = {
+    try {
+      logger.info(s"loading configure")
+      ConfigSource
+        .fromConfig(
+          ConfigFactory.load().getConfig("DBConfig") // just want this piece of the config file
+        )
+        .load[DBConfig]
+        .left
+        .map(errors => ConfigurationException(errors.toList.map(f => f.toString)))
+        .map { config =>
+          logger.info(s"config: $config")
+          logger.info(s"url: ${config.url}")
+          // todo enable connection pool, compare this code with Database.forConfig()
+          val factory = new DaoFactory(Database.forURL(config.url))
+          logger.info(s"got factory")
+          factory
+        }
+    } catch {
+      case t: Throwable => Left(ConfigurationException(List(t.getMessage)))
+    }
+  }
+
+  def shutdown(database: Database): Try[Unit] = Try {
+    database.close()
+  }.recoverWith { case t: Throwable =>
+    Failure(ConfigurationException(List(t.getMessage), Some(t)))
+  }
+
+}
